Split single WorkerIterableSourceWorker into separate workers
Rather than having a single worker implementation which supports all the different types of data sources, this change splits into separate worker implementations for each data source.

Since each worker needs to property wrap and expose the source functions to the main thread. For this the WorkerIterableSourceWorker is re-purposed to provide this wrapper for each worker implementation to use to wrap their source implementations.